backend/requirements.txt
Flask==3.1.2
python-dotenv==1.2.1

Flask
flask-cors
psycopg2-binary
python-dotenv
